AIQuantum/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pydantic>=1.8.0
pyyaml>=5.4.1

# Machine Learning
scikit-learn>=0.24.0
tensorflow>=2.6.0
torch>=1.9.0

# Data and Trading
ccxt>=1.60.0
ta-lib>=0.4.0
python-binance>=1.0.0

# Visualization and Dashboard
rich>=10.0.0
plotly>=5.3.0
dash>=2.0.0

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910 